requirements.txt
antlr4-python3-runtime==4.9.3
boto3==1.42.73
botocore==1.42.73
hydra-core==1.3.2
jmespath==1.1.0
omegaconf==2.3.0
packaging==26.0
psutil==7.2.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3
